add checking if file path exists before renaming

+++ Src/DuckTagsFolderStructureManager.py
@@ -24,7 +24,9 @@ class DuckTagsFolderStructureManager(object):
 
         if self.__convert_file_name_with_pattern__(file_path, file_format_pattern_index):
             new_file_path = self.__join_file_path__()
-            os.rename(file_path, new_file_path)
+
+            if not os.path.exists(new_file_path):
+                os.rename(file_path, new_file_path)
 
     def __slice_file_path__(self, file_path):
         file_path_root_length = file_path.rfind(r'/') + 1
